# Fewer wire allocs in SQL() implementations