--- a/sql/expression/function/inet_convert.go
+++ b/sql/expression/function/inet_convert.go
@@ -81,7 +81,7 @@ func (i *InetAton) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
 	}
 
 	// Expect to receive an IP address, so convert val into string
-	ipstr, err := types.ConvertToString(val, types.LongText)
+	ipstr, err := types.ConvertToString(val, types.LongText, nil)
 	if err != nil {
 		return nil, sql.ErrInvalidType.New(reflect.TypeOf(val).String())
 	}