--- a/sql/types/datetime.go
+++ b/sql/types/datetime.go
@@ -411,7 +411,6 @@ func (t datetimeType) SQL(_ *sql.Context, dest []byte, v interface{}) (sqltypes.
 		return sqltypes.Value{}, sql.ErrInvalidBaseType.New(t.baseType.String(), "datetime")
 	}
 
-	//valBytes := AppendAndSliceBytes(dest, val)
 	valBytes := val[start:]
 
 	return sqltypes.MakeTrusted(typ, valBytes), nil