--- a/sql/types/enum.go
+++ b/sql/types/enum.go
@@ -49,6 +49,7 @@ var (
 type EnumType struct {
 	collation             sql.CollationID
 	hashedValToIndex      map[uint64]int
+	valToIdx              map[string]int
 	indexToVal            []string
 	maxResponseByteLength uint32
 }
@@ -70,7 +71,8 @@ func CreateEnumType(values []string, collation sql.CollationID) (sql.EnumType, e
 	// including accounting for multibyte character representations.
 	var maxResponseByteLength uint32
 	maxCharLength := collation.Collation().CharacterSet.MaxLength()
-	valToIndex := make(map[uint64]int)
+	hashedValToIndex := make(map[uint64]int)
+	valToIdx := make(map[string]int)
 	for i, value := range values {
 		if !collation.Equals(sql.Collation_binary) {
 			// Trailing spaces are automatically deleted from ENUM member values in the table definition when a table
@@ -82,11 +84,12 @@ func CreateEnumType(values []string, collation sql.CollationID) (sql.EnumType, e
 		if err != nil {
 			return nil, err
 		}
-		if _, ok := valToIndex[hashedVal]; ok {
+		if _, ok := hashedValToIndex[hashedVal]; ok {
 			return nil, fmt.Errorf("duplicate entry: %v", value)
 		}
 		// The elements listed in the column specification are assigned index numbers, beginning with 1.
-		valToIndex[hashedVal] = i + 1
+		hashedValToIndex[hashedVal] = i + 1
+		valToIdx[value] = i + 1
 
 		byteLength := uint32(utf8.RuneCountInString(value) * int(maxCharLength))
 		if byteLength > maxResponseByteLength {
@@ -95,8 +98,9 @@ func CreateEnumType(values []string, collation sql.CollationID) (sql.EnumType, e
 	}
 	return EnumType{
 		collation:             collation,
-		hashedValToIndex:      valToIndex,
+		hashedValToIndex:      hashedValToIndex,
 		indexToVal:            values,
+		valToIdx:              valToIdx,
 		maxResponseByteLength: maxResponseByteLength,
 	}, nil
 }
@@ -309,6 +313,9 @@ func (t EnumType) Collation() sql.CollationID {
 
 // IndexOf implements EnumType interface.
 func (t EnumType) IndexOf(v string) int {
+	if idx, ok := t.valToIdx[v]; ok {
+		return idx
+	}
 	hashedVal, err := t.collation.HashToUint(v)
 	if err == nil {
 		if index, ok := t.hashedValToIndex[hashedVal]; ok {